Add logic after drive

+++ BE/drivemate/src/main/java/DriveMate/drivemate/domain/SemiRouteLineString.java
@@ -17,7 +17,7 @@ public class SemiRouteLineString extends SemiRoute{
     private String name;
     private String description;
     private Integer distance; // Integer 인가?
-    private Integer time; // Integer 인가?
+    private Double time; // Integer 인가? <--- 여기서 오류 날수도 있어요
     private Integer roadType;
     private Integer facilityType;
 }
